refactor(browser-common): share browser config types
